+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Lupa change log
 ===============
 
+Latest changes
+--------------
+
+* new module function ``lupa.lua_type()`` that returns the Lua type of
+  a wrapped object as string, or ``None`` for normal Python objects.
+
+
 1.0.1 (2014-10-11)
 ------------------